Added a parent (Entity) class to the existing character inheritance hierarchy because I want to create a EntityViewModel class and a character view model class.
