+++ src/main/java/UI/presenters/CharacterViewModel.java
@@ -1,0 +1,4 @@
+package UI.presenters;
+
+public class CharacterViewModel {
+}
